More consistent method of keeping final documents
Instead of not adding them to the tree, we should do as we do for other cases
(newer than TeX file, read only, etc), and choose not to delete it at the same
time we consider all the other files. This also has the advantage of having
output consistent with the rest of the program.